+++ NEWS.md
@@ -2,6 +2,8 @@
 * The `logger_name` is used in myClim object instead of the `logger_index`.
 * The `logger_name` is visible in `mc_plot_line`.
 * The `show_logger_name` parameter has been added to the `mc_reshape_wide` function.
+* The `serias_name` label changed to `series` in `mc_plot_line`.
+* The progress bar is added to `mc_prep_crop`.
 
 # myClim 1.3.3
 * A performance issue in the `mc_prep_crop` function has been fixed.